Wip2.7

+++ wip/a/get_next_line.h
@@ -9,10 +9,10 @@
 #  define BUFFER_SIZE 1024;
 # endif
 
-char	*ft_strdup(const char *s1);
-int	ft_strlen(const char *str);
-char	*ft_strjoin(char const *s1, char const *s2);
-char	*ft_substr(char const *s, int start, int len);
+char    *ft_strdup(const char *s1);
+size_t ft_strlen(const char *str);
+char    *ft_strjoin(char const *s1, char const *s2);
+char    *ft_substr(char const *s, int start, int len);
 char *get_next_line(int fd);
 
 #endif
